Manual report creation

+++ app/domain/models/evidence.py
@@ -5,7 +5,7 @@ class Evidence(db.Model):
     __tablename__ = "evidences"
 
     id = db.Column(db.Integer, primary_key=True)
-    report_id = db.Column(db.Integer, db.ForeignKey("reports.id"), nullable=False)
+    report_id = db.Column(db.BigInteger, db.ForeignKey("reports.id"), nullable=False)
     file_name = db.Column(db.String(255), nullable=False)
     file_path = db.Column(db.String(500), nullable=False)
     mime_type = db.Column(db.String(120), nullable=True)
